feat: introduce contextual color for lyne-title and lyne-journey-header which enables to style lyne-footer contextually
Introducing contextual colors has been implemented with
the usage of `pseudo-private custom properties`. We use the
dash-dash-underline pattern for internal, so called `pseudo-private
custom properties`. We expose the same custom property without
the underline in the variable name and map it to the internal
variable. Within this step we also assign a default value
as fallback if the exposed custom property was not defined.