fix(sbb-tag): fix accessibility by using buttons instead of checkboxes (#1647)
As the `sbb-tag` can be used with a radio behavior, checkbox behavior or even combined, we changed the way we built the `sbb-tag`. Now there are buttons and corresponding aria behavior which can be used in both cases. There is a new property called `multiple` to enable multiple mode (checkbox) or using it with exclusive mode (radio). The exclusive mode is the new default.

BREAKING CHANGE:
 - Exclusive mode is now the default mode. To change to multiple, set the new `multiple` property to true.
 - You should always set value property of the `sbb-tag` to work correctly in the `sbb-tag-group`.