fix(sbb-alert, sbb-teaser, sbb-teaser-hero): use paragraph for content slot due to semantic reasons (#1785)
BREAKING CHANGE:

Internally we wrapped slotted content with a `<p>` tag in sbb-alert, sbb-teaser and sbb-teaser-hero to guarantee semantic meaning of descriptions. This should not be a problem as long there are no other semantic elements slotted. Anyways, basically only textual elements are intended to be slotted by design. 

Please check your usages:
 - `sbb-alert`: default slot
 - `sbb-teaser`: `description` slot
 - `sbb-teaser-hero:`: default slot