fix(sbb-form-field): fix accessibility when slotting form error (#1922)
Previously, aria-describedby of the input was not correctly updated when removing or first adding a sbb-form-error. Additionally, to support removing and adding sbb-form-error multiple times for screen readers, we had to move the aria-live region to a single element, although according to documentation the aria-live region should be on a container.