fix: fix scrollbar styles for Chrome
Chrome started to support scrollbar-width and scrollbar-color CSS properties
which we are currently using as a fallback for Firefox. However, when using it in Chrome,
it disables all the -webkit-scrollbar stylings. With the new supports check we can avoid to
switch to the scrollbar-color approach in Chrome as long as the old approach still works.
Painpoint: the new approach doesn't allow a different color for hover.