fix(sbb-dialog): fix accessibility with option to hide the header on scroll (#2231)
In order to support the new feature of optionally hiding the header during scroll, we had to re structure all the dialog component.

BREAKING CHANGE: The `sbb-dialog` component now supports the dedicated inner elements `sbb-dialog-title`, `sbb-dialog-content`, and `sbb-dialog-actions`. Use these components to respectively provide a title, a content and, optionally, a footer with an action group. Moreover, the full-screen variant (which occurred when no title was provided to the dialog) has been removed. To achieve a full-screen overlay, please use the `sbb-overlay` component.
This was the previous implementation:
```html
<sbb-dialog title-content="Title">
  <p>Dialog content.</p>
  <sbb-action-group slot="action-group">...</sbb-action-group>
</sbb-dialog>
```
This is the new implementation:
```html
<sbb-dialog>
  <sbb-dialog-title>Title</sbb-dialog-title>
  <sbb-dialog-content>Dialog content</sbb-dialog-content>
  <sbb-dialog-actions>...</sbb-dialog-actions>
</sbb-dialog>
```
Previously, a ***full-screen*** dialog was displayed if no title was provided to the dialog component:
```html
<sbb-dialog>
  <p>Dialog content.</p>
</sbb-dialog>
```
It is now mandatory to use the `sbb-overlay` component to display the same variant:
```html
<sbb-overlay>
  <p>Overlay content.</p>
</sbb-overlay>
```