build: implement visual regression testing with @web/test-runner (#2624)
This PR sets up the basic configuration for visual regression testing with `@web/test-runner`.

`web-test-runner.config.js` contains a new test group, which is only activated when explicitly called with `yarn test:visual-regression`. This test group introduces a new file pattern `*.snapshot.spec.ts`, which contains the visual regression tests.
Additionally, there is a new GitHub Actions workflow, which runs the visual regression tests and saves failed screenshots in an artifact. This artifact is then downloaded in the secure workflow, which adds an `in_progress` check, if any failed screenshots exist.

We do not want to save the snapshot png in the repository, as this would bloat it considerably. Due to this, we use containers as an external storage of the png files. The process works as follows:
Each continuous integration run on our main branch creates/updates a container `baseline` (https://github.com/lyne-design-system/lyne-components/pkgs/container/lyne-components%2Fvisual-regression). For any other CI run (in the secure context) this container is used as a service and the png files are downloaded when needed.
Additionally, the baseline image is hosted at `https://lyne-visual-regression-baseline.app.sbb.ch/` in order to enable local visual regression testing.

As for local visual regression testing; To avoid diffs between operating system browser differences, visual regression testing is only directly run in a Linux environment. If it is executed in any other OS, it will try to run it in a container to prevent OS differences showing up in the snapshots.

---------

Co-authored-by: Jeremias Peier <jeremias.peier@sbb.ch>