fix: attempt to mitigate SWC minifier
The problem with the SWC minifier seems to be caused by using isBrowser() in the static getter on SbbLanguageController. We replace this usage by isServer from lit, which is a constant.
This should prevent the symbol from isBrowser() to be removed, as the compiler can already remove it during build time.