refactor(sbb-tab): align with sbb-stepper (#2744)
BREAKING CHANGE: the `sbb-tab-title` component has been renamed to `sbb-tab-label`. 
A new component named `sbb-tab` has been created, and it is now the only supported tag for the tab's content; `article`, `section` and `div` are not supported anymore. 
To solve the issue #1351, the `didChange` event on `sbb-tab-group` now includes an object which contains the currently selected tab index, the `sbb-tab-label` component and related `sbb-tab` component, plus, if available, the previous ones.