feat(sbb-train-formation): introduce new types and refactoring (#3199)
Closes #2681

BREAKING CHANGE: The `hide-wagen-label` property of the
`sbb-train-formation` was removed. Now it automatically doesn't show the
label if no label is set on all the wagons.
The i18n `i18nClosedCompartmentLabel()` method doesn't take
`wagonNumber` as an argument anymore but is a constant now.
Additionally, there are some visual changes:
- `sbb-train-wagon`: The `ouccpancy` property doesn't default to `none`
anymore but to `null`. Please replace the currently undefined occupancy
property with the value `none`.
- `sbb-train-wagon`: Previously for the locomotive the label was not
displayed, but now it would, as soon as there is one provided
- `sbb-train-formation`: The inline padding (left / right) was removed
but can be set by CSS variable. See documentation.