fix(sbb-alert, sbb-alert-group): remove dismissal event and method (#3216)
BREAKING CHANGE: The deprecated `dismissalRequested` event and
`requestDismissal()` method of `sbb-alert` have been removed. The
`sbb-alert` handles its closing and DOM removal on his own. If the
closing should be prevented, the `willClose` event can be canceled. The
`didDismissAlert` of the `sbb-alert-group` has been removed. As
alternative, consumers can listen to the `didClose` event of an
`sbb-alert`.