refactor(sbb-header): removed 'logo' slot (#3230)
Closes #3093 

BREAKING CHANGE: removed the `logo` slot and the default `sbb-logo` from
the `sbb-header`.
Alternatively, add the `sbb-header-logo` CSS class to the logo or to an `<a>`-element
containing the logo and use the default slot. To align the logo to the right, use a spacer element
with the `sbb-header-spacer` CSS class applied. For more information, see `sbb-header` docs.