fix: avoid moving content when blocking scrolling
With previously introduced changes to scroll block,
in certain situation the content is slightly moving.
This happens when there is a scroll bar visible and
 some content is aligned next to the scrollbar.
Setting the height to 100dvh ensures that the
body is always taking the full available height
 which is necessary e.g. for the sidebar.
Setting position: relative reverts the previous
change where this was removed.