fix: provide `accessibility-current` property for link elements
With the previous implementation, the `aria-current` was set on the host
(e.g. in breadcrumb or navigation). However, this was not forwarded to the
inner anchor tag which resulted in partially broken screen reader support.
This PR introduces the `accessibility-current` property on link elements which
forwards the value to the inner anchor element as `aria-current`.