fix: synchronize id references (#3505)
We have several components that rely on an id connection. Before this
change, when the id was not found when setting the id, the referenced
element was never found. With this change we introduce an id reference
observer which allows finding id's in the document (rootNode) whenever
they appear or vanish. This imitates the browser behavior for id
connections.