feat: introduce possibility of complex values on form elements (#3676)
Closes https://github.com/sbb-design-systems/lyne-angular/issues/111
Closes https://github.com/sbb-design-systems/lyne-angular/issues/76
Closes https://github.com/sbb-design-systems/lyne-angular/issues/73
Closes https://github.com/sbb-design-systems/lyne-angular/issues/54

BREAKING CHANGE: We had to change some default values and types in order to support complex values for our form components. The value property of the SbbFormAssociatedMixin is now abstract.
- For sbb-autocomplete-grid-option, sbb-chip, sbb-option and
sbb-toggle-option, the default value is now null instead of an empty
string
- For button-like components, sbb-date-input and sbb-time-input, the
default value is now an empty string instead of null

---------

Co-authored-by: Tommmaso Menga <tommaso.menga-94@hotmail.it>