fix(sbb-toast): replace property `dismissible` with `readOnly` and change logic
BREAKING CHANGE: Removed property `dismissable` and replaced with `readOnly`
of the `sbb-toast`. The `sbb-toast` now displays the close button per default.
To hide the close button, use the `readOnly` property. Also, the timeout has been
changed to `0`, which means that a toast is not closing automatically per default.
Furthermore, the close button can now also be displayed beside an action,
it is not replacing it anymore.