fix(sbb-toggle): change elements dimensions to fit intrinsic size (#3794)
Closes #3788 

BREAKING CHANGE: Before this change, the intrinsic size of the
`sbb-toggle` was smaller than the element itself. Now the height and
width of the element has increased to fit the real dimensions (plus 2px
on all sides). Potential action needed: For some layouts the
`sbb-toggle` needs a negative margin of 2px in order to nicely fit in
existing layouts (align background track of `sbb-toggle` to the rest of
the content).