refactor: rework datepicker implementation (#3828)
This PR refactors and restructures the datepicker and related
components.

BREAKING CHANGE: The `<sbb-datepicker>` and related components have been
fundamentally refactored.
The native `<input>` element is no longer supported and the
`<sbb-date-input>` is now the main control the following components
connect to: `<sbb-datepicker>`, `<sbb-datepicker-toggle>`,
`<sbb-datepicker-previous-day>` and `<sbb-datepicker-next-day>`
The `<sbb-datepicker>` and `<sbb-datepicker-toggle>` had their API
surface reduced, as most is now controlled/configured via
`<sbb-date-input>`. This includes properties, methods and events.
The `now` property has been removed from the `<sbb-datepicker>` and
`<sbb-calendar>`. If you need to set the current date for testing
purposes, you can stub the `today()` method of the `defaultDateAdapter`,
which is used in the background.
In the `<sbb-form-field>` the `<sbb-datepicker-toggle>`,
`<sbb-datepicker-previous-day>` and `<sbb-datepicker-next-day>`
components are now slotted according to their position to
`<sbb-date-input>` in the DOM.