fix(sbb-calendar): avoid inconsistent label when aborting selection (#4311)
Closes #4308

When the calendar is loaded, the day view shows a month (or two in wide)
based on the currently calculated 'week' array, which is based on the
activeDate.
On the other hand, the month label on top is calculated based again on
the activeDate.

When the year / month is changed, the new day view is reloaded using a
temporary activeDate, which is calculated based on the chosen values for
year and month and the same day of the 'real' active date.

Since this date is temporary, the value is not saved; if users change
again to year/month view and then exit as discussed in #4308, the month
label is recalculated based on the 'real' active date (meaning: the
selected, or today).

To fix this misalignment, the 'init' method is called when the view is
reset, so the view always matches the active date.